requirements.txt
glfw==2.7.0
imgui==2.0.0
multipledispatch==1.0.0
ndspy==4.2.0
numpy==2.1.1
PyOpenGL==3.1.7
pyrr==0.10.3
